requirements.txt
Django==4.1.7
django-debug-toolbar==4.1.0
djangorestframework==3.14.0
colorama==0.4.6
click==8.1.6
python-dotenv==1.0.0
setuptools==65.5.1
xhtml2pdf==0.2.16
django-import-export
djangorestframework-datatables-editor